requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
alembic==1.7.6
    # via -r requirements.in
anyio==3.5.0
    # via starlette
asgiref==3.5.0
    # via uvicorn
click==8.0.4
    # via uvicorn
ecdsa==0.17.0
    # via python-jose
fastapi==0.74.1
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
h11==0.13.0
    # via uvicorn
idna==3.3
    # via anyio
importlib-metadata==4.11.1
    # via alembic
importlib-resources==5.4.0
    # via alembic
mako==1.1.6
    # via alembic
markupsafe==2.1.0
    # via mako
mysql-connector-python==8.0.28
    # via -r requirements.in
passlib==1.7.4
    # via -r requirements.in
protobuf==3.19.4
    # via mysql-connector-python
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pydantic==1.9.0
    # via fastapi
python-dotenv==0.19.2
    # via -r requirements.in
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
rsa==4.8
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   python-multipart
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.31
    # via
    #   -r requirements.in
    #   alembic
starlette==0.17.1
    # via fastapi
typing-extensions==4.1.1
    # via pydantic
uvicorn==0.17.5
    # via -r requirements.in
zipp==3.7.0
    # via
    #   importlib-metadata
    #   importlib-resources
